// MMA301/Code/cinema-app-fe/src/navigator/type.ts
export enum AppRoutes {
  LOGIN = "LOGIN",
  REGISTER = "REGISTER",
  RESET_PASSWORD = "RESET_PASSWORD",
  TABS = "TABS",
  USER_INFORMATION = "USER_INFORMATION",
  INFO_FILM_DETAIL = "INFO_FILM_DETAIL", 
  SCHEDULE_FILM_DETAIL = "SCHEDULE_FILM_DETAIL",
  TIME_SCHEDULE_FILM_DETAIL = "TIME_SCHEDULE_FILM_DETAIL",
  ROOM_FILM = "ROOM_FILM",
  PAYMENT = "PAYMENT",
  HISTORY_TICKET = "HISTORY_TICKET",
}